// src/payload-types.ts
export interface Recipe {
  id: number;
  title: string;
  /**
   * Auto-generated from title when left empty.
   */
  slug: string;
  heroImage?: (number | null) | Media;
  description?: string | null;
  categories?: (number | Category)[] | null;
  prepMinutes?: number | null;
  cookMinutes?: number | null;
  servings?: number | null;
  difficulty?: ('easy' | 'medium' | 'hard') | null;
  ingredients?:
    | {
        quantity?: string | null;
        unit?: string | null;
        item: string;
        note?: string | null;
        id?: string | null;
      }[]
    | null;
  instructions?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  /**
   * Per serving.
   */
  nutrition?: {
    calories?: number | null;
    /**
     * grams
     */
    protein?: number | null;
    /**
     * grams
     */
    netCarbs?: number | null;
    /**
     * grams
     */
    fat?: number | null;
    /**
     * grams
     */
    fiber?: number | null;
  };
  publishedAt?: string | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    hero?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    og?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number;
  name: string;
  /**
   * Auto-generated from name when left empty.
   */
  slug: string;
  updatedAt: string;
  createdAt: string;
}